Register

+++ app/controllers/admin_controls_controller.rb
@@ -3,9 +3,25 @@ class AdminControlsController < ApplicationController
   before_action :set_admin_control , only: [:show,:edit,:destroy,:update]
   include AdminControlsHelper
   include ComboBoxHelper
+  include RegisterBooksHelper
 
   def index
-    @admin_controls = AdminControl.all
+    @responsible_functions = ResponsibleFunction.where(enable:true).map{|x| [x.name, x.id]}
+    @search = if params[:search]
+                Date.parse(params[:search])
+              else
+                Date.today
+              end
+
+
+
+    filters = {}
+    @responsible = params[:responsible]
+    if @responsible  && @responsible != "" then
+      filters["responsible_function_id"] = @responsible
+    end
+    @admin_controls = AdminControl.all.by_month(@search).where(filters).paginate(page: params[:page], per_page: 30).order('created_at desc')
+    @months = (Date.today - 1.year..Date.today).map(&:beginning_of_month).uniq.reverse.map{|dt| [dt.strftime("%Y-%m"), dt] }
   end
 
 
@@ -41,7 +57,12 @@ class AdminControlsController < ApplicationController
     combo_box_data()
     @admin_control = AdminControl.new(admin_control_params)
     @admin_control.user_id = current_user.id
+
       if @admin_control.save
+        @admin_control.sa =  "sa-#{@admin_control.id}"
+        @admin_control.save
+        create_register_admin(@admin_control,current_user)
+        # up_register =  RegisterBook.where(admin_control_id: @admin_control.id).update(code_sa: @admin_control.sa)
          redirect_to @admin_control, notice: "Admin control was successfully created."
       else
         combo_box_data()
@@ -54,6 +75,8 @@ class AdminControlsController < ApplicationController
   def update
        combo_box_data()
       if @admin_control.update(admin_control_params)
+        puts "#{current_user.id} #{current_user.email}"
+        update_register_admin_control(@admin_control,current_user)
         redirect_to @admin_control, notice: "Admin control was successfully updated."
       else
 
@@ -65,10 +88,72 @@ class AdminControlsController < ApplicationController
 
 
   def destroy
-    if @admin_control.destroy
-       redirect_to admin_controls_url, notice: "Admin control was successfully destroyed."
+    checklist = Checklist.where(admin_control_id: @admin_control.id)
+    checklist_items = ChecklistItem.where(checklist_id: checklist.pluck(:id) ).delete_all
+    checklist_destroy = Checklist.where(admin_control_id: @admin_control.id).delete_all
+    if destroy_register_admin_control(@admin_control,current_user)
+        if delete_ref_admin_control(@admin_control)
+           if @admin_control.destroy
+              redirect_to admin_controls_url, notice: "Admin control was successfully destroyed."
+           end
+        end
     end
+
   end
+
+
+
+  def create_register_admin(obj,current_user)
+    register = RegisterBook.new(admin_control_id:obj.id,
+                                create_user:"#{current_user.first_name}" + " " + "#{current_user.last_name}",
+                                create_time:Date.today,offer_id: nil,
+                                concept_admin: obj.concept.name,theme_admin:obj.theme.name,
+                                type_admin:obj.type.name,item_admin: obj.type.name,company: obj.company.name,
+                                code_sa: obj.sa,code: obj.code,rol_responsible: obj.responsible_function.name,is_admin_control:true )
+    register.save
+  end
+
+
+  #Time.now.strftime("%I:%M:%S %P")
+  def update_register_admin_control(obj,current_user)
+    RegisterBook.where(admin_control_id:obj.id).update(update_user:"#{current_user.first_name}" + " " + "#{current_user.last_name}",
+                                                       update_time:Date.today,
+                                                       concept_admin: obj.concept.name,
+                                                       theme_admin:obj.theme.name,
+                                                       type_admin:obj.type.name,
+                                                       item_admin: obj.item.name,
+                                                       rol_responsible: obj.responsible_function.name,
+                                                       code_sa: obj.sa,code: obj.code,
+                                                       company:obj.company.name)
+  end
+
+
+
+
+  def destroy_register_admin_control(obj,current_user)
+    @result = ""
+    dia_actual = Date.today
+    register = RegisterBook.where(admin_control_id:obj.id)
+                   .update(destroy_user:"#{current_user.first_name}" + " " + "#{current_user.last_name}",
+                           destroy_time: dia_actual,time_destroy:Time.now.strftime("%I:%M:%S %P"))
+    if register
+      @result = true
+    else
+      @result = false
+    end
+    return @result
+  end
+
+  def delete_ref_admin_control(obj)
+    r = RegisterBook.where(admin_control_id: obj.id).update(admin_control_id: nil)
+    if r
+      result_admin = true
+    else
+      result_admin = false
+    end
+    return result_admin
+  end
+
 
   private
